pavo/requirements.txt
ffmpeg-python==0.2.0
numpy==1.24.1
# opencv-python==4.7.0.68
pika==1.3.1
wget==3.2